Refactor code into multiple files
